CRUDVersion

+++ spring-project/src/main/java/com/example/springproject/entity/Event.java
@@ -3,8 +3,6 @@ package com.example.springproject.entity;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
-import com.example.springproject.entity.User;
-import com.example.springproject.entity.Student;
 
 import java.util.List;
 
@@ -14,14 +12,13 @@ import java.util.List;
 public class Event {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
     private String name;
     private String date;
     private String description;
     private String category;
 
     @OneToMany
-//    @JoinColumn(name = "student_id")
     private List<Student> students;
 
 }
